Moved `debug_pane` into it's separate module
