+++ components/monitors/cu_consolemon/src/lib.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "debug_pane")]
+mod debug_pane;
 pub mod sysinfo;
 
 use ansi_to_tui::IntoText;
@@ -31,16 +33,8 @@ use std::time::Duration;
 use std::{io, thread};
 use tui_nodes::{Connection, NodeGraph, NodeLayout};
 use tui_widgets::scrollview::{ScrollView, ScrollViewState};
-
 #[cfg(feature = "debug_pane")]
-use {
-    compact_str::CompactStringExt,
-    log::{Level, LevelFilter, Log, Metadata, Record},
-    std::collections::VecDeque,
-    std::io::Read,
-    std::sync::atomic::AtomicU16,
-    std::sync::mpsc::{Receiver, SyncSender},
-};
+use {debug_pane::UIExt, log::Log};
 
 #[cfg(feature = "debug_pane")]
 const MENU_CONTENT: &str = "   [1] SysInfo  [2] DAG  [3] Latencies  [4] Debug Output  [q] Quit   ";
@@ -301,100 +295,7 @@ struct UI {
     #[cfg(feature = "debug_pane")]
     error_redirect: gag::BufferRedirect,
     #[cfg(feature = "debug_pane")]
-    debug_output: DebugLog,
-}
-
-#[cfg(feature = "debug_pane")]
-struct DebugLog {
-    debug_log: VecDeque<String>,
-    max_rows: AtomicU16,
-    rx: Receiver<String>,
-}
-
-#[cfg(feature = "debug_pane")]
-impl DebugLog {
-    fn new(max_lines: u16) -> (Self, SyncSender<String>) {
-        let (tx, rx) = std::sync::mpsc::sync_channel(1000);
-        (
-            Self {
-                debug_log: VecDeque::new(),
-                max_rows: AtomicU16::new(max_lines),
-                rx,
-            },
-            tx,
-        )
-    }
-
-    fn push_logs(&mut self, logs: String) {
-        if logs.is_empty() {
-            return;
-        }
-
-        self.debug_log.push_back(logs);
-        let max_row = self.max_rows.load(Ordering::SeqCst) as usize;
-        while self.debug_log.len() > max_row {
-            self.debug_log.pop_front();
-        }
-    }
-
-    fn update_logs(&mut self) {
-        let max_row = self.max_rows.load(Ordering::SeqCst) as usize;
-
-        for log in self.rx.try_iter() {
-            if log.is_empty() {
-                continue;
-            }
-
-            self.debug_log.push_back(log);
-            if self.debug_log.len() > max_row {
-                self.debug_log.pop_front();
-            }
-        }
-    }
-
-    fn get_logs(&mut self) -> String {
-        self.update_logs();
-        self.debug_log.join_compact("").to_string()
-    }
-}
-
-#[cfg(feature = "debug_pane")]
-#[derive(Clone)]
-struct LogSubscriber {
-    tx: SyncSender<String>,
-}
-
-#[cfg(feature = "debug_pane")]
-impl LogSubscriber {
-    fn new(tx: SyncSender<String>) -> Self {
-        let log_subscriber = Self { tx };
-        log::set_boxed_logger(Box::new(log_subscriber.clone())).unwrap();
-        log::set_max_level(LevelFilter::Info);
-        log_subscriber
-    }
-
-    fn push_logs(&self, log: String) {
-        if let Err(err) = self.tx.send(log) {
-            eprintln!("Error Sending Logs to MPSC Channel: {}", err.0)
-        }
-    }
-}
-
-#[cfg(feature = "debug_pane")]
-impl Log for LogSubscriber {
-    fn enabled(&self, metadata: &Metadata) -> bool {
-        metadata.level() <= Level::Debug
-    }
-
-    fn log(&self, record: &Record) {
-        if self.enabled(record.metadata()) {
-            let message = format!("[{}] - {}\n", record.level(), record.args());
-
-            self.push_logs(message);
-        }
-    }
-
-    fn flush(&self) {}
+    debug_output: debug_pane::DebugLog,
 }
 
 impl UI {
@@ -405,7 +306,7 @@ impl UI {
         task_stats: Arc<Mutex<TaskStats>>,
         task_statuses: Arc<Mutex<Vec<TaskStatus>>>,
         error_redirect: gag::BufferRedirect,
-        debug_output: DebugLog,
+        debug_output: debug_pane::DebugLog,
     ) -> UI {
         init_error_hooks();
         let nodes_scrollable_widget_state =
@@ -570,25 +471,6 @@ impl UI {
         )
     }
 
-    #[cfg(feature = "debug_pane")]
-    fn draw_debug_output(&mut self, f: &mut Frame, area: Rect) {
-        let mut error_buffer = String::new();
-        self.error_redirect
-            .read_to_string(&mut error_buffer)
-            .unwrap();
-        self.debug_output.push_logs(error_buffer);
-
-        let debug_output = self.debug_output.get_logs();
-
-        let p = Paragraph::new(debug_output).block(
-            Block::default()
-                .title(" Debug Output ")
-                .title_bottom(format!("{} log entries", self.debug_output.debug_log.len()))
-                .borders(Borders::ALL),
-        );
-        f.render_widget(p, area);
-    }
-
     fn draw(&mut self, f: &mut Frame) {
         let layout = Layout::default()
             .direction(Direction::Vertical)
@@ -743,7 +625,7 @@ impl CuMonitor for CuConsoleMon {
             #[cfg(feature = "debug_pane")]
             {
                 let max_lines = terminal.size().unwrap().height - 5;
-                let (debug_log, tx) = DebugLog::new(max_lines);
+                let (debug_log, tx) = debug_pane::DebugLog::new(max_lines);
 
                 // redirect stderr, so it doesn't pop in the terminal
                 let error_redirect = gag::BufferRedirect::stderr().unwrap();
@@ -758,7 +640,7 @@ impl CuMonitor for CuConsoleMon {
                 );
 
                 #[allow(unused_variables)]
-                let log_subscriber = LogSubscriber::new(tx);
+                let log_subscriber = debug_pane::LogSubscriber::new(tx);
 
                 // Override the cu29-log-runtime Log Subscriber
                 #[cfg(debug_assertions)]
